flask-crud-api/requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
paho-mqtt
